refactor(egg): change find_team_egg to static and move hatch_egg_for_client to player.c
